refactor: move IFW rule UI components and mappers to sub-package
- Relocate `GlobalIfwRuleEditorComponents.kt`, `GlobalIfwRuleUiComponents.kt`, `GlobalIfwRuleMapper.kt`, `SimpleGlobalIfwRuleEditorComponents.kt`, and `GlobalIfwRuleResources.kt` to the `com.merxury.blocker.feature.globalifwrule.impl.components` package
- Update package declarations and import statements across the module to reflect the new structure
- Update copyright year to 2026 in `GlobalIfwRuleEditorComponents.kt`
- Minor formatting adjustment in `GlobalIfwRuleViewModel.kt`
