+++ feature/globalifwrule/impl/src/main/kotlin/com/merxury/blocker/feature/globalifwrule/impl/AddRuleScreen.kt
@@ -50,6 +50,10 @@ import com.merxury.blocker.core.ui.previewparameter.AdvancedGlobalIfwRuleScreenP
 import com.merxury.blocker.core.ui.previewparameter.AdvancedGlobalIfwRuleScreenPreviewState
 import com.merxury.blocker.core.ui.previewparameter.GlobalIfwRulePreviewParameterData
 import com.merxury.blocker.feature.globalifwrule.api.R
+import com.merxury.blocker.feature.globalifwrule.impl.components.ComponentTypeDropdown
+import com.merxury.blocker.feature.globalifwrule.impl.components.IntentFilterBanner
+import com.merxury.blocker.feature.globalifwrule.impl.components.RuleEditorTopBar
+import com.merxury.blocker.feature.globalifwrule.impl.components.SwitchRow
 import com.merxury.core.ifw.editor.IfwEditorNode
 import com.merxury.core.ifw.editor.hasTopLevelComponentFilter
 import com.merxury.core.ifw.model.IfwComponentType
