+++ feature/globalifwrule/impl/src/test/kotlin/com/merxury/blocker/feature/globalifwrule/impl/GlobalIfwRuleMapperTest.kt
@@ -20,6 +20,10 @@ import com.merxury.blocker.core.model.data.AdvancedGlobalIfwRuleDraft
 import com.merxury.blocker.core.model.data.GlobalIfwRuleEditMode
 import com.merxury.blocker.core.model.data.SimpleGlobalIfwRuleDraft
 import com.merxury.blocker.core.model.data.SimpleTargetMode
+import com.merxury.blocker.feature.globalifwrule.impl.components.toIfwRule
+import com.merxury.blocker.feature.globalifwrule.impl.components.toIfwRuleOrNull
+import com.merxury.blocker.feature.globalifwrule.impl.components.toRuleItemUiState
+import com.merxury.blocker.feature.globalifwrule.impl.components.toSimpleDraftOrNull
 import com.merxury.core.ifw.editor.IfwEditorConditionKind
 import com.merxury.core.ifw.editor.IfwEditorNode
 import com.merxury.core.ifw.editor.IfwEditorStringMatcherMode
